# frontend/openapi.json
{
	"components": {
		"schemas": {
...
			"CreateMCPOAuthAuthorizationInputBody": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/CreateMCPOAuthAuthorizationInputBody.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"approved": {
						"description": "Whether the user approved the MCP OAuth request",
						"type": "boolean"
					},
					"client_id": {
						"description": "OAuth client ID or client metadata URL",
						"type": "string"
					},
					"code_challenge": {
						"description": "PKCE S256 code challenge. Required when approved is true.",
						"type": "string"
					},
					"code_challenge_method": {
						"description": "PKCE method. Must be S256 when approved is true.",
						"type": "string"
					},
					"redirect_uri": {
						"description": "OAuth redirect URI",
						"type": "string"
					},
					"resource": {
						"description": "MCP protected resource URL",
						"type": "string"
					},
					"response_type": {
						"description": "OAuth response type. Must be code.",
						"type": "string"
					},
					"scope": {
						"description": "Requested OAuth scope. Defaults to mcp:full.",
						"type": "string"
					},
					"state": {
						"description": "Opaque client state to echo to the redirect URI",
						"type": "string"
					},
					"workspace_id": {
						"description": "Optional workspace ID the resulting MCP token is limited to",
						"type": "string"
					}
				},
				"required": ["approved", "response_type", "client_id", "redirect_uri"],
				"type": "object"
			},
			"CreateMCPOAuthAuthorizationOutputBody": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/CreateMCPOAuthAuthorizationOutputBody.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"redirect_url": {
						"description": "URL the browser should redirect to after authorization",
						"type": "string"
					}
				},
				"required": ["redirect_url"],
				"type": "object"
			},
...
			"ErrorDetail": {
				"additionalProperties": false,
				"properties": {
					"location": {
						"description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
						"type": "string"
					},
					"message": {
						"description": "Error message text",
						"type": "string"
					},
					"value": {
						"description": "The value at the given location"
					}
				},
				"type": "object"
			},
			"ErrorModel": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/ErrorModel.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"detail": {
						"description": "A human-readable explanation specific to this occurrence of the problem.",
						"examples": ["Property foo is required but is missing."],
						"type": "string"
					},
					"errors": {
						"description": "Optional list of individual error details",
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": ["array", "null"]
					},
					"instance": {
						"description": "A URI reference that identifies the specific occurrence of the problem.",
						"examples": ["https://example.com/error-log/abc123"],
						"format": "uri",
						"type": "string"
					},
					"status": {
						"description": "HTTP status code",
						"examples": [400],
						"format": "int64",
						"type": "integer"
					},
					"title": {
						"description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
						"examples": ["Bad Request"],
						"type": "string"
					},
					"type": {
						"default": "about:blank",
						"description": "A URI reference to human-readable documentation for the error.",
						"examples": ["https://example.com/errors/example"],
						"format": "uri",
						"type": "string"
					}
				},
				"type": "object"
			},
...
		}
	},
	"info": {
		"title": "OpenPost API",
		"version": "1.0.0"
	},
	"openapi": "3.1.0",
	"paths": {
...
		"/mcp/oauth/authorize": {
			"post": {
				"operationId": "create-mcp-oauth-authorization",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMCPOAuthAuthorizationInputBody"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateMCPOAuthAuthorizationOutputBody"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Unauthorized"
					},
					"403": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Forbidden"
					},
					"422": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Unprocessable Entity"
					},
					"500": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"summary": "Create or deny an MCP OAuth authorization response",
				"tags": ["Auth"]
			}
		},
...
	}
}